Q220-013 no info about termination when no implicit contract
When a non-terminating function is called from an assertion or from
a contract, GNATprove emits an info message stating that its
(implicit) contract will not be available for proof.
When there is no explicit contract and the return type does not
need a dynamic invariant, there is nothing to ignore. We detect
this case to not emit the info message.

* gnat2why-util.ads
(Type_Needs_Dynamic_Invariant): Functions that return True if
function result of a given type need additional information (in the
form of a dynamic type invariant).

* gnat2why/why/gnat2why-expr.adb
(Transform_Function_Call): Do not emit messages about implicit
contract being ignored if there is no implicit contract.

Change-Id: Ie62cc7125bc378f22298e6c99af31bef6aec7dde