Q222-023 add a minimized reproducer
This is a minimal reproducer for a problem with a call chain that has
SPARK_Mode On->Off->On->Off and ultimately accesses a Part_Of variable of
abstract state, whose full refinement is not known to the tool (because of a
SPARK_Mode => Off on the body of package with abstract state).

* testsuite/gnatprove/tests/Q222-023__flow_part_of_spec/
Add; disabled, because currently crashes.

Change-Id: I459f6a7b562746349811c3a72e7cefbf3f0580ef