P805-023 Change classwide handling for top-down marking
Move some code for the handling of class-wide types from the declaration
to the entity. In top-down marking, the declaration of a type might not be
processed, but its entity will.

* spark_definition.adb
(Mark): removed code to update classwide mapping
(Mark_Entity): added code to update classwide mapping, and mark full view
  of a tagged type

Testsuite impact: some extra messages on record definitions that are not
in SPARK are removed. But I checked that the actual violation (usually
related to discriminants on derived types) is still properly reported.

Change-Id: I7b026d8187f9f826a082b98b48d9e903111803e3