Q222-023 Entities_To_Translate should only include those in SPARK
Having entities "not in SPARK" in Entities_To_Translate is not only
counterintuitive; it also takes memory and requires extra filtering based on
Entity_In_SPARK later (which takes extra CPU time).

This patch cleans this up in preparation for detecting references to missing
globals in the gnat2why itself (currently they are detected by Why3 as
crashes due to imports of undeclared modules).

* gnat2why/spark/spark_definition.adb
(Mark_Entity): only append entity to Entities_To_Translate if it is "in
 SPARK", i.e. has no SPARK violations.

* gnat2why/why/gnat2why-driver.adb
(Translate_Entity): spec: add a precondition (an imperfect one, but better
 than nothing); body: don't check SPARK status of the entity.

(Do_Generate_VCs): assert and do not check SPARK status of types.

(Translate_Standard_Entity): ignore standard entities that are not in SPARK,
 e.g. double-precision floating type.

(Is_Translated_Subprogram): now only called for entities in SPARK (which
 in ensured by a ghost code and erased in production).

(Complete_Declaration): same as above.

* gnat2why/why/gnat2why-tables.adb
(Store_Information_For_Entity): this procedure is now only called on types
 in SPARK.

Change-Id: Iacfd3bde1e7dd47e33421c3b63139cfa60ba30fd