Q301-007 use epsilon to protect axioms for function contracts
When a function has an inconsistent contract (a contract which
cannot hold for some inputs), GNATprove generates an unsound
axiom which may then allow to prove anything in a caller of such
a function, and so, even if the function is always called on
'valid' inputs, that is, inputs on which the contract holds.
Though this behavior is not unsound, and is theoretically expected
with a proof technology such as SPARK, it appears to come as a
surprise to some users.
In this patch, we avoid generating unsound axioms as much as
possible by introducing guards for function axioms which are only
assumed to hold on actually used values. Note that there are still
cases where an unsound axiom will be generated (functions called in
type invariants / type predicates, in primitive equalities of
record types, or sometimes in user written quantified expressions).
As a consequence, having inconsistent contracts on functions is
still a bad usage of SPARK, and users should be instructed to avoid
it.
Also, this new 'safer' translation can sometimes impact proof
capabilities. Thus, we provide an advanced switch --no-axiom-guard
to disable it.
This switch impacts generation of Why3 code. Bost versions of the
generation should be tested enough though as the default generation
'includes' the specific one.

* common/gnat2why_args.adb
(Proof_Generate_Guards_Name): Name of the new switch.
(Proof_Generate_Guards): Value of the no axiom guard switch.
(Write_To_File): Update the value of the Proof_Generate_Guards
 variable.

* gnat2why/why/gnat2why-expr-loops.adb
(Transform_Loop_Statement): For loops iterating over a container,
 use an epsilon for calls to has_element/element when necessary.

* gnat2why/why/gnat2why-expr.adb
(Transform_Function_Call): Use an epsilon when necessary.
(Transform_Quantified_Expression): For quantification over a
 container, use an epsilon for calls to has_element/element when
 necessary.

* gnat2why/why/gnat2why-subprograms.adb
(Translate_Subprogram_Spec): Declare a predicate to act as a guard
 in the axioms for the function.
(Generate_Axiom_For_Post): Add a guard to the axiom for the post of
 functions.
(Generate_Subprogram_Program_Fun): Assume that the guard predicate
 holds in the post of program functions.
(Translate_Expression_Function_Body): Add a guard to the axiom for
 the definition of expression functions. Do not generate an axiom
 for functions inlined for proof.

* gnat2why/why/gnat2why-types.adb
(Generate_Type_Completion): Use an epsilon for definition of the user
provided primitive equality over records whe necessary.

* gnat2why/why/gnat2why-util.adb
(Use_Guard_For_Function): Return true if axioms for a function
 contracts should be prefixed by a guard.

* gnat2why/why/why-atree-modules.adb
(Insert_Subprogram_Symbols): Insert symbols for guard predicates.

* gnat2why/why/why-atree-sprint.adb
(Print_Epsilon): Print new epsilon construct.

* gnat2why/why/why-gen-expr.adb
(New_Function_Call): Translate a function call either directly or
 using an epsilon. Functions are translated using an epsilon except
 if either they are predefined, they are inlined for proof, or they
 are defined in a package with external axioms.

* gnat2why/why/why-gen-names.adb
(Why_Name_Enum): Add names for post predicates.

* gnat2why/why/why-sinfo.ads
(W_Epsilon): New why3 node for epsilon construct.

* gnatprove/configuration.adb
(No_Axiom_Guard): Switch for disabling generation of guards
 completely.

* gnatprove/gnatprove.adb
(Pass_Extra_Options_To_Gnat2why): Inform gnat2why when the
 generation of guards is disabled.

* share/spark/help.txt
 Add new switch in help message.

Change-Id: Ic9f62f748e4222405b04d4fa7fa951d7e1c84b3f