Q308-010 Clean Check_Elaborate_Body
The body of Check_Elaborate_Body does not need to cope with non library
unit packages.

* gnat2why/flow/flow-analysis.adb
(Check_Elaborate_Body): removed code that copes with non library unit
packages and removed decisions on whether to do the check (they are now
made before the call in flow.adb).

* gnat2why/flow/flow.adb
(Flow_Analyse_CUnit): extend conditions to decide when to do the check

Change-Id: I4800bc5bac3998eaac811a2ec4c04ab0980bce0b