Q312-002 use sequences in the ring buffer test in adacore_u
As noted on this TN, the test does not go through anymore with the
built-in necoding of arrays. Modify it to use sequences instead of
arrays to regain the proofs.
The test is not exactly the same as the one presented in the sildes
anymore but the explanations stay the same.

Change-Id: Ie55e1aba0c2f1b89768b0914b0949e3cd0313433