Q313-012 ignore unreachable "others" case branches
Case branches with empty "others" choices are now detected and marked as
exceptional. The effect on the control-flow graph is the same as putting a
"raise" statement in the list of choice statements.

* gnat2why/spark/spark_util.ads
* gnat2why/spark/spark_util.adb
(Is_Empty_Others): new routine to detect empty "others" choices.

* gnat2why/flow/flow-control_flow_graph.adb
(Do_Case_Statement): mark unreachable CASE branch as abnormal.

* testsuite/gnatprove/tests/Q313-012__case_statement/test.out Output
Updated; spurious check due to lack of initialization in unreachable branch
is gone.

Change-Id: I5a64020f4a98ed131a9962febd5bb0d482617afa