packages with no external axioms don't go into Entity_Set
Marked entities are included in Entity_Set to prevent repeated marking when
they are referenced many times. This is useless for packages, because they
just can't be referenced from anywhere. However, they were still included in
this set, because flow analysis and proof wrongly iterated it.

After recent fixes, flow and proof iterates over Entity_List (except for one
place in flow analysis, but it is most likely a bug and it looks for record
types and not packages anyway). There is no need to include packages in
Entity_Set anymore.

Note: this doesn't apply to packages with external axioms, which need to be
guarded against repeated marking. Most likely this ensures that a package
with external axioms goes into Entity_List before any of its entities and
that it appears there no more than once.

(no-tn-check)

* gnat2why/spark/spark_definition.adb
(Mark_Package_Declaration): don't include ordinary packages in Entity_Set.

Change-Id: I1a2201c18965eae5c77997d51e5cc7c3dff8c7ef