Q320-013 use hashed sets for marking results
In translation to Why the Retysp routine and its queries to marking results
can be a performance hotspot, because the results are kept in the "default"
ordered Node_Sets. Since we never iterate over these sets the ordering does
not matter anyway. Thus in marking it is safe and faster to use hashed sets.

* gnat2why/utils/common_containers.ads
(Hashed_Node_Sets): new container type.

* gnat2why/spark/spark_definition.ads
* gnat2why/spark/spark_definition.adb
Switch from ordered to hashed sets where possible.

Change-Id: I5913b56e64797ca941552df9586413465636c134