minor avoid identity conversion of Flow_Id
We were changing Flow_Id variant to Normal_Use for Flow_Ids coming both from
Get_Globals, which is fine, because their variants are In_View and Out_View;
and from Direct_Mapping_Id, which is useless, because it is already
Normal_Use.

This is now fixed by a simple code move.

* gnat2why/flow/flow_utility.adb
(Get_Depends): avoid useless calls to Change_Variant.

Change-Id: I5bcf6e628bf34b816623dee37d0b698d75a4a8be