Q328-033 add argument for tag of dispatching functions
The return value of a dispatching function call will depend on the
tag of the controlling argument, which may not be, in the case of a
controlling result, a parameter of the function.
To avoid assuming in Why that two dispatching calls with the same
parameters are always equal, add a new parameter to dispatching
functions for the tag of the controlling argument.

* gnat2why/why/gnat2why-expr.adb
(Transform_Function_Call): Give the tag of the controlling argument
 as an additional argument to the function when the call is
 dispatching.

* gnat2why/why/gnat2why-subprograms.adb
(Tag_Binder): Binder to be used as additional tag argument for
 dispatching functions.
(Generate_Axiom_For_Post): Adapt to new tag argument of dispatching
 functions.
(Generate_Subprogram_Program_Fun): Dispatching program functions
 take an additional parameter for the dispatching tag.
(Translate_Subprogram_Spec): Dispatching logic functions take an
 additional parameter for the dispatching tag.

Change-Id: Ia22f47fb1011f67da389dd2f7cefe341e3a75751