store mapping from Entity_Name to Symbol in vector
Mapping from Entity_Name to their Symbol representation (which is really a
unique string) was stored in hashed map. This was inefficient, because
Entity_Name are continuous. Now we keep them in Vector.

Note: we could use GNAT.Table instead of Vector, because it grows linearly,
but its API interface is less elegant.

* gnat2why/utils/common_containers.adb
(String_Cache): rewrite from hashed map to vector.

(no-tn-check)

Change-Id: I4bb7ab02aa324802a8705f9d3487e5a9f71c3248