Q406-018 only check private types whose enclosing unit is in SPARK
We emit checks for default values of private types. These checks
should only be emitted if the enclosing unit of the type is
explicitely marked as SPARK_Mode On, and not if it is Auto.

* gnat2why/why/gnat2why-driver.adb
(Do_Generate_VCs): Test whether types are declared in a unit with
spec in SPARK before generating VCs for them.

Change-Id: I565e96409a0f018dd32f11c322e118c4292ee0fc