P909-031 reliable locations for tasking checks
Tasking checks always refer to an object (e.g. unsynchronized variable or
entry of a protected object) and to one or more objects with task
components. However, in the worst case none of these is known by Entity_Id
and their Entity_Names may depend on platform (e.g. because the code "withs"
platform-specific Ada runtime units).

For reliable error messages this patch introduces a preference to locate
checks on:
1) task objects from the current compilation unit (sorted by Entity_Id)
2) task objects from other compilation units (sorted by Entity_Id)
3) task objects known by Entity_Name (sorted alphabetically)
4) object, if it is known by Entity_Id
5) the root of the current compilation unit
and sorts the continuations messages about conflicting tasks using the same
preference.

This is a substantial rewrite, because in the meantime the code was
duplicated to implement entry-wise checks for multiple entries of a single
protected object (for GNAT Extended Ravenscar profile). Also, it removes few
inefficiencies in that code.

* gnat2why/flow/flow_generated_globals-phase_2.ads
* gnat2why/flow/flow_generated_globals-phase_2.adb
(<): ordering relation for task objects.
(Task_Instances_Maps): change from lists to ordered multi-sets.

* gnat2why/flow/flow-analysis.adb
(Check_Concurrent_Accesses): substantial rewrite.

* testsuite/gnatprove/tests/*/test.out
Update slocs and ordering of tasking-related checks.

Change-Id: I8802e6e8834074ef4d966121a64516cab8edd4f8