Q327-024 Forbid plain precondition on dispatching subprograms
The new rule SPARK RM 6.1.1(2) forbids the use of plain preconditions
(typically with aspect Pre) on dispatching subprograms, which should
instead use Pre'Class. The reason for that is that otherwise the
dynamic semantics of Ada combined with the verification of Liskov
Substitution Principle in SPARK force Pre and Pre'Class to be
equivalent. Hence it would be useless to have both. Note that it is
still possible to have Pre on a primitive operation of an untagged
private type, as there is no way to dispatch on such a subprogram in
SPARK (dispatching on this subprogram is forbidden, and deriving such
a type is also forbidden).

* gnat2why/spark/spark_definition.adb
(Mark_Violation): Add possibility for continuation message.
(Mark_Subprogram_Entity): Detect violation of new rule.

* gnat2why/spark/spark_util-subprograms.adb
(Filter_Classwide_Contracts): Fix for classwide contracts not from
 an aspect, when GNAT pragmas Pre_Class and Post_Class are used.

* testsuite/gnatprove/tests/K505-010__stack/stacks.ads
* testsuite/gnatprove/tests/K505-010__stack/test.out
* testsuite/gnatprove/tests/N709-001__contracts/test.out
* testsuite/gnatprove/tests/N709-001__contracts/unit1.ads
* testsuite/gnatprove/tests/N709-001__dispatch/object.ads
* testsuite/gnatprove/tests/N709-001__dispatch/test.out
* testsuite/gnatprove/tests/N715-015__flow_oo_legal/professionals.ads
* testsuite/gnatprove/tests/N715-015__flow_oo_legal/test.out
* testsuite/gnatprove/tests/O309-037__LSP_dynamic_type/lsp_checks.adb
* testsuite/gnatprove/tests/O512-022__stacks/stacks.ads
* testsuite/gnatprove/tests/O512-022__stacks/test.out
* testsuite/gnatprove/tests/O901-014__formal_vectors/aida-....ads
* testsuite/gnatprove/tests/O901-014__formal_vectors/test.out
* testsuite/gnatprove/tests/P718-001__ubstr_good/aida-...ptr-mutable.ads
* testsuite/gnatprove/tests/P718-001__ubstr_good/aida-...shared_ptr.ads
* testsuite/gnatprove/tests/P718-001__ubstr_good/test.out
* testsuite/gnatprove/tests/Q327-014__no_pre/dispatch.ads
* testsuite/gnatprove/tests/Q327-014__no_pre/test.out
Adapt tests and outputs, to use Pre'Class instead of Pre when needed.

* testsuite/gnatprove/tests/Q327-024__classwide_pre/test.opt
* testsuite/gnatprove/tests/Q327-024__classwide_pre/test.out
Test passes now.

Change-Id: I871db5f6a6fe15fc13b098daa6af6adddd753bce