P708-032 Added support for counterexamples with z3
This commit adds support for counterexamples with z3. An option is added
to run-test.py. Gnatprove help is not updated because this flag is not
officially supported for customers.

* gnatprove/configuration.adb
(Postprocess): Added Z3_Counterexample switch.

* gnatprove/gnatprove.adb
(Prepare_Why3_Manual): Added flag --ce_prover with option z3_ce to change
the prover for counterexamples to z3_ce for gnatwhy3 when flag
z3-counterexample is on.

* share/spark/config/why3.conf
(z3_ce): Added a version of z3 with driver for counterexamples.

* testsuite/gnatprove/lib/python/test_gnatprove.py
(run_testsuite): Added option z3_counterexample for internal testing.

* testsuite/gnatprove/lib/python/test_support.py
(gnatprove): Added option --z3-counterexample

* testsuite/gnatprove/tests/P708-032__riposte_basic_z3_CE/*
(test): Adding a test for Z3 generation of counterexamples.

Change-Id: I397d9cb02342bf3b812899f8f650eaea845b4a84