Q418-008 fix usage of RTS option with --codepeer
We need to pass the --RTS info to the codepeer_gprbuild. To be able to
do that, we also need to add that option to the spark_codepeer_wrapper,
and all pass the option when invoking it.

Since we only pass the exact same option around in
spark_codepeer_wrapper, we don't even take the time to properly parse it
and just store the --RTS=<path> commandline argument as a whole.

* gnatprove.adb
(Run_Codepeer): pass --RTS option to spark_codepeer_wrapper if needed
* spark_codepeer_wrapper:
(Parse_Command_Line): parse --RTS option
(Generate_Scil): add --RTS option to codepeer-gprbuild if present

Change-Id: I081e3bedbea09014581f030ec0ffd96bdca47aaf