Q407-005 avoid conversions on selected components
When a record type contains discriminant dependant components, it
can be the case that the Etype of a selected component does not
match the type of of the component in the corresponding object.
When transforming selected components, use the type of the
component in the prefix as the expected type to avoid unnecessary
conversions.

* gnat2why/why/gnat2why-expr.adb
(Bind_From_Mapping_In_Expr): Introduce a conversion to the
 Type_Of_Node of the expression to be bound.
(Insert_Ref_Context): When fetching the actual, use the
 Type_Of_Node as the expected type.
(Transform_Expr): For selected components, use the type of the why
 field access to avoid conversions.

Change-Id: Id15968f323b8eea6fea56dd8be7d5ec05a8bd8aa