Q418-004 translate powers of 2 as shifts on modulars
The power function is not supplied by SMTlib on bitvectors and
hardly axiomatized in Why3 theory. To improve provability, rather
translate powers of 2 on modulars into shifts, at least when the
modulus is a power of two.

* gnat2why/why/gnat2why-expr.adb
(Transform_Expr): Special case powers of two on modular types with
binary modulus and translate them as shifts.

Change-Id: I382d61d4666fcf4d854bf191e1452a86ef25db2c