Q420-007 use mathematical integers to compute power
On modular types with non-binary modulus, applications of the power
operator cannot be translated on bitvector types as it will
introduce an inappropriate modulus in case of wraparound. Instead,
translate them as applications of power on mathematical integers,
followed by an appropriate modulo.

* gnat2why/why/gnat2why-expr.adb
(New_Op_Expr): Do not use the power operator on the base bitvector
 type when the modulus is not a power of two. Rather call
 Transform_Non_Binary_Modular_Operation to get a specialized
 handling. When the modulus is a power of two, perform a modulo
 after the call to power when necessary.
(Transform_Non_Binary_Modular_Operation): Translate a call to the
power operator on mathematical integers to avoid wraparounds and
perform a modulo afterward to convert back to the modulor type.

Change-Id: I1ff6d409b9fa3511dbb4fc544d4202f65a2fc58a