P805-023 preregister entities that are registered globally anyway
Some symbols are registered globally (that is, never removed again) in the
symbol map used for proof. We do this in one go at the beginning instead
of one-by-one.

This removes an ordering constraint on the processing of subprogram
entities in SPARK, and will simplify top-down marking.

* gnat2why-driver.adb
(Register_Symbol): new function to register a symbol in the symbol map
(Translate_CUnit): add preprocessing of subprogram symbols
* gnat2why-subprograms.adb
(Translate_Subprogram_Spec): remove code that adds subprogram entities to
  symbol map
* testsuite: minor proof changes, probably due to ordering of declarations

Change-Id: I2416ca8162a3366af197731de61d73413df4e5ec