remove dead code for "Aspect => foo" dependency map
Dependency maps, i.e. Depends and Initializes aspects, written as

  "Aspect => foo"

are translated to "Aspect => (foo)" by the frontend, i.e. their expressions
are translated to N_Aggregate nodes. The code that expected N_Identifier was
never executed, and was wrong anyway (because it should take the Entity of
the N_Identifier, not the identifier itself).

* gnat2why/flow/flow_dependency_maps.adb
(Parse_Raw_Dependency_Map): remove dead code.

(no-tn-check)

Change-Id: I4c771627e34b6d6afbd6f3d7ea269187684d52b4