PB07-039 Consider null records in control flow graph
We did not have any vertex for null records. This resulted in wrong control
flow graphs and ultimately in a crash for the test with this ticket number.
We now have this vertex as a result of a modification in Flatten_Variable.

* gnat2why/flow/flow-analysis.adb
(Find_Ineffective_Imports_And_Unused_Objects): check if the type is a null
record before emitting messages. Also detect if we are dealing with a type
predicate and in this case emit a different message.
(Find_Use_Of_Uninitialized_Variables): check if the type is a null record
before emitting messages.

* gnat2why/flow/flow-analysis.adb
(Flatten_Variable): if the type is a null record it returns the variable
itself.

* testsuite/gnatprove/tests
Update several test outputs. Some come from considering null records and
some from having a better message in case of type predicates.

Change-Id: Ia90dbe16289185ec2241673e366958a65e96c94d