Q502-017 Use unique names for private parts of record types
For types with private parts not visible in SPARK, we used to
generate a component of type __private to stand for the invisible
components. Unfortunately, it was causing troubles when using
interactive provers, as it was not possible to map these private
parts to different types. Also, it was theoretically wrong, as
it implied that primitive equality on these types were the same.
I was not able to derive an unsoundness from this using
automatic provers though...

* gnat2why/why/gnat2why-tables.adb
(Has_Private_Part): Return True if a type has its own private
 part (not inherited from its parent).

* gnat2why/why/why-atree-modules.adb
(Insert_Type_Symbol): Register symbols for unique type and equality
 of private parts of a type which has a private part.

* gnat2why/why/why-gen-names.adb
(Why_Name_Enum): Add names for type and equality functions on
 private parts.

* gnat2why/why/why-gen-records.adb
(W_Type_Of_Component): Return the appropriate abstract type for
 private parts.
(New_Field_Equality): Return the appropriate abstract function for
 private parts.
(Declare_Rep_Record_Type): For types with private parts, declare an
 abstract type and an uninterpreted equality function for the
 private parts.

* gnat2why/why/why-inter.adb
(Eq_Base): Support being called on EW_Abstract types with no ada
 node standing for abstract why types. In this case, compare the
 names instead of comparing the ada nodes.

Change-Id: Iee3bc94f9f134dbcced4f2575bba4c29d7cffe06