Q221-039 Allow access to constants in SPARK
This is the first addition of support for access types in SPARK.
For now, only access to constant types are supported, and values of
such a type can only be created by taking the address (through attribute
Access) of an aliased constant which has no variable input. Thus, no
information flow through such values only depend on assignment to the
access values, not the assignments to accessed values. Verification
conditions should be generated to check that dereferenced access values
are not null.

* docs/lrm/source/declarations-and-types.rst
Authorize named access to constant types. Refine the definition of
constants with/without variable inputs to state that this only applies
to stand-alone constants.

* docs/lrm/source/language-defined-aspects-and-attributes.rst
Authorize Access attribute.

* docs/lrm/source/lexical-elements.rst
Authorize keyword "access".

* docs/lrm/source/names-and-expressions.rst
Authorize attribute references with Access on aliased constants
without variable input.

Change-Id: Id6216c8c4764ab1307505ec4a38b0ab25340f09e