Q502-022 fix dependency computation of Why3 record modules
the automatic computation of dependencies for axiom theories was broken,
because of the indirection of the "rep" theory that is created for
records.

If no "defining entity" is specified when closing the theory, then the
dependencies are not computed correctly. So the fix here is to specify
that defining entity when closing the theory for a rep theory for
records.

why-gen-records.adb:
(Create_Rep_Record_Theory_If_Needed): specify the record type entity as
the defining entity when closing the theory

Change-Id: Idb28eb07c2dc855c967b027dd6fbd7aa5e2e343c