P118-034 fixes in preparation for the main merge
Fixes that should really belong to master, but apparently they only apply to
a super-obscure code that is not even worth to write a unit test. The new gg
we would crash without them, because it is less forgiving on data coming
from the frontend globals.

* gnat2why/flow/flow_generated_globals-phase_1.adb
(GG_Register_Direct_Calls): assertion was too restrictive for default values
 of function parameters to generic units.

* gnat2why/spark/spark_register.adb
(Process_Node): register calls to inlined functions that return array
 aggregates or integer literals.

Change-Id: I53007094e23a5cf92190943049bcbe8c43b4939b