Q525-025 assume value of tag of function results
We used to not assume anything about the tag of the results of
function calls. This didn't have a huge proof impact because once
the result of the function is stored in an object, the tag will be
determined by the type of the object. However, it was not
prossible to prove that the result of a function call was equal to
the value of the object in which it was stored (we did not know
that the tags had to match).
We now add an imlicit postcondition to functions returning specific
tagged types that the tag of the result is the tag of the type. For
dispatching functions with controlling result, we rather use the
additional tag parameter as the tag of the result will depend on
dynamic dispatching.

* gnat2why/why/gnat2why-subprograms.adb
(Generate_Axiom_For_Post): Assume the value of the tag of tagged
 results.

Change-Id: I29b2f88c560789cc973c8d5bbcdabce5589fc85c