Q529-009 Avoid crash on use of incorrect tagged primitives
In proof, we assumed that overriding of tagged primitives could not
access more gloabl variables than their ancestor. However, since
this analysis is only done by flow analyzing the corresponding
function declaration, we can still go to proof on users of such
subprograms if they are declared in distinct units.
The code is now more robust and does not crash in this situation
anymore.

* gnat2why/why/gnat2why-subprograms.adb
(Generate_Dispatch_Compatibility_Axioms): Do not generate axioms
 for compatibility of overriding failing the global compatibility
 checks.

Change-Id: I4b1366d6b8815c6b69fb045f1d46c064bf6548df