Q522-012 Int to float conversion should use RNE rounding mode
Ada RM prescribes the rounding used on the machine for int to float
conversions. In SPARK, Round-to-Nearest-ties-to-Even is used, hence
int to float conversions should use RNE rounding mode. Currently RNA
was used erroneously.

* gnat2why/why/why-atree-modules.adb
Use correct name for rounding.

* share/spark/theories/_gnatprove_standard.mlw
Fix names and modes for float rounding.

* testsuite/gnatprove/tests/PB16-052__float_invariant/test.out
Update test with small proof regressions, some of which can be recovered
with higher steps.

Change-Id: Iaa133383dbfe16a4659d4d21a1514261a515b8db