Q529-038 fix infinite loop during marking of pragma Annotate
A pragma annotate on a body triggers also the insertion of the sloc
range of the spec, usually also including following declarations that
are logically part of the spec declaration (e.g. predicate functions).
This will loop in the case of a pragma annotate on a type with a
predicate.

The reason is that in the case of a predicate, the predicate function
spec and body are both part of the same declaration list, and are both
marked as not coming from source. So when the body is included, the spec
will be included as well, including its following declarations that are
not from source, which also includes the body. Infinite loop.

gnat2why-annotate.adb:
(Insert_With_Next): extra argument to not insert a specific node
(Insert_Annotate_Range): when inserting the range of the spec for some
body, exclude the body from being inserted again

Change-Id: Id04d53ee35bf0e528b3c168d87d6914c7bed7403