Q515-026 Added a missing if in printing of record counterexamples
This commit adds a condition which checks that counterexamples types are
correct with respect to their entity. This condition existed for all types
except for records (which is a mistake).
As there are transformations in the Spark to cvc4 pass and that cvc4 may
not give every counterex or that we may not parse them, this condition is
necessary. And there is no other easy known way to do it.

* gnat2why/why/gnat2why-counterexamples.adb
(Refine): Adding a case checking that counterexamples of record type
actually correspond to entities of record type.
(Replace_Question_Mark): Function that replaces empty string with question
mark for better printing in records.

* testsuite/gnatprove/tests/Q515-026__sk
(tests): Changed output.

* testsuite/gnatprove/tests/*
(tests): Update tests.

Change-Id: I430bec0f24b46fd4c92ff1a33d9b91ae68a003f7