M607-025 tune ordering of provers
With the default ordering of provers on this test (*):
1) alt-ergo runs for  ~90sec/8GB and discharges no VCs
2) cvc4     runs for ~150sec/1GB and discharges most of the VCs
3) z3       almost immediately discharges few remaining VCs

Performance data from a pretty powerful desktop; on a less powerful (but
still decent) laptop it causes swapping and makes the OS generally
unresponsive.

It seems better to run z3 first, as it discharges all the VC in a several
seconds and needs no more than 1MB of RAM.

* testsuite/gnatprove/tests/M607-025__rev_stack/test.py
Change ordering of provers.

* testsuite/gnatprove/tests/M607-025__rev_stack/test.out
Update output: all VCs are now proved by Z3.

Change-Id: I7b207021ac062734009045fac0c407298acd3869