PB07-037 remove entries in connection map when joining them
Routine Join combines Nodes (an IN parameter) into a Block (an OUT
parameter) adding a new entry into a CM (i.e. a "connection map", an IN OUT
parameter). Now it also removes entries of the combined Nodes from the map,
which is in the spirit of abstracting and never looking back into details.

This primarily forces a cleaner code where this routine is used, especially
in Do_Package_Body_Or_Stub which is quite difficult to understand.

The extra benefit is that by calling Join after processing any list of nodes
the size of the connection map is reduced, which helps to keep it small and
thus fast to use.

This patch doesn't change the semantics itself, but prepares the code for
more serious work in the context of several TNs related to generation of the
Initializes contract.

* gnat2why/flow/flow-control_flow_graph.adb
(Join): remove joined entries from the connection map.

(Do_Iterator_Loop): connections for the loop itself are now removed by the
 call to Join and cannot be updated in-place, but need to be re-inserted.

(Do_Package_Body_Or_Stub): update connections by referring to the abstracted
 block.

Change-Id: I5fefc69c772b1913a921cde9a5ce5bac73fc9467