PB07-037 move instead of copy connections
Coping connections within a connection map can be replaced with a more
efficient moving. An extra benefit is that the old entry is removed from the
connection map, so future accesses to the map are faster and its callees are
saner.

In particular, this only affects Do_Package_Body_Or_Stub, which is known for
being notoriously hard to understand (and relying on connection map being
copied contributed to this).

No change in semantics, but this is done in preparation for several TNs that
touch generation of the Initializes aspect.

* gnat2why/flow/flow-control_flow_graph.adb
(Move_Connections): rewritten from Copy_Connections.

(Do_Package_Body_Or_Stub): reference the rewritten connections, no the
 original ones.

Change-Id: I43e2cbe3f067415eaa26ae3d2ff602953f39140d