Q613-001 add test with combination of task's Depends contract
Task's dependency on itself is implicit, but might be explicitly confirmed
in the Depends clause. Same for single task objects. This test have all
combinations: single/ordinary task type, explicit/implicit dependency. All
should be accepted in the flow analysis.

* testsuite/gnatprove/tests/Q613-001__flow_task_depends/
Add; the current flow errors are wrong, the code is fine as far as the flow
analysis is concerned.

Change-Id: Ic3be44c1481bc149a4589cd4c72a747c74491434