Q531-032 check discriminant on Unchecked_Union
We were relying on the frontend flag Do_Discriminant_Check to
decide whether or not a discriminant check was needed for a record
access. This flag is not set for access to types with pragma
Unchecked_Union as discriminant checks are not required for them
at runtime, so we were not emitting a check in GNATprove either.
We now emit a discriminant check for accesses to components
whenever they are part of a record variant part.

* docs/lrm/source/language-defined-pragmas.rst
Unchecked_Union is supported.

* gnat2why/spark/spark_definition.adb
(Mark_Pragma) Do not warn about ignored pragma on Unchecked_Union.

* gnat2why/why/gnat2why-tables.adb
(Find_Rec_Node_For_Variant): Return the entity to be used to query
 the variant information for a given type.
(Get_Variant_Info): Use Find_Rec_Node_For_Variant.
(Has_Variant_Info): Check whether a record component is located in
 a variant part.

* gnat2why/why/why-gen-records.adb
(New_Ada_Record_Access): Generate a discriminant check for every
 component access inside a variant part.

Change-Id: Ia7c876c6b68d8743d11902703d608fa692bdce56