only detect variable inputs in user-defined equality of subprograms
Move code for detecting variable inputs in user-defined equality out of
traversal (where it was executed also for tasks and packages). Now this code
is only executed for subprograms. No change in semantics, only small change
of the location of the message (from "function" to "=").

(no-tn-check)

* gnat2why/flow/flow-analysis-sanity.adb
(Check_Expressions): move code.

Change-Id: If74ce19cd5106bf2b819d932b4a98648f0802b23