Q622-010 Missing range check on static scalar derived types
GNATprove was not emitting range checks for scalar derived types
when the bounds were static. This is incorrect when the type which
is derived itself has dynamic bounds because then the frontend
won't check the bounds statically and we will have a hole.
Emit the check if either the type itself or its Etype has dynamic
bounds.

* gnat2why/why/gnat2why-expr.adb
(Transform_Declaration): Emit a range check on scalar derived types
 and subtypes if either the type itself or its Etype has dynamic
 bounds.

Change-Id: I8848503f44c8ce0d5a954a04d1f4609dc508aae7