Q623-006 crash on hidden discriminant dependent components
When a record component is hidden by a private declaration, the
frontend may remove it from the list of components of type
entities. In GNATprove, we still want to be able to access it, so
we retieve hidden components from ancestor types and add them to
the list of component of the derived type.
If a hidden component depends on a discriminant, then it will refer
to the entity for the discriminant in the ancestor type, not in the
current type, leading to a crash in the translation because this
entity is not stored in the symbol table.
We normalize entities stored in the symbol table so that there is
only one entity per discriminant of a type hierarchy (the one from
the rrot type).

* gnat2why/why/gnat2why-util.adb
(Normalize_Entity): On discriminants, return the associated
 discriminant in the root record type.
(Ada_Ent_To_Why): Normalize enties before insertion and lookup into
 the table.

* gnat2why/why/why-gen-records.adb
(Build_Predicate_For_Record): No need to insert discriminants of
 the base type into the symbol table anymore (it was done to
 translate default values of component types).

Change-Id: I8409512909fa45384659f95ed5a7287e4b3ebf92